ZKL tiny minor patch: prevent a rare ZKL crash if user attempted to scroll an empty webpage.
Note: this is a really low significant because by default internal browser already preloaded everything, But it could happen on old ZKL if external link is clicked & user attempted to scroll. To replicate this rare crash (for demo): disable auto-login (so that it doesn't preload page) and attempt to scroll (using scroll-wheel) on the main (empty) page.

svn path=/trunk/other/; revision=10425