max-width for "Last battles" and "Posted threads"
Added a maximum width for the div elements that include the last ten battles played and the last ten threads that were posted in by the player. This is to stop "Posted threads" from being forced down and create a lot of empty space if a mapname of thread title is very long.

Only has an effect in the default skin where the parent's width is known, not in the alternative skin used by ZKL that uses the full window.