Update Discord.net to 3.2.0
The goal is to see if updating the Discord library will fix the flakiness on the bridge since the current library version is a few years old and Discord has modified their keepalive logic since then. I've verified that the basic ChatRelay works on my machine on my Discord server, but cannot test full operation with more than two users, so this PR remains fairly risk to merge.

I've moved the Discord dependency from manual compilation to a package in the hopes of making this process easier in the future since this involved few other dependency updates (mainly Newton.Json) and adds relatively little risk.