Improve parsing of infolog_full.txt
The new parser identifies regions of the infolog that correspond
to individual games. Each game is further parsed to find:
  GameID
  FirstDesyncIndex
  GameState file names

If the infolog contains GameState file names,
include them in a table in the Issue description.

If the infolog contains multiple games with desyncs, try
to include all of the desync messages in the truncated
infolog.